# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NazLibra POS API",
        "description": "Offline-first REST API for NazLibra POS mobile application. All protected routes require Bearer token + X-Tenant-Slug header.",
        "contact": {
            "email": "support@omnevo.net"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
    "paths": {
...
        "/api/v1/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get POS dashboard KPIs for a date range",
                "description": "GET /api/v1/dashboard?from=<date>&to=<date>\n\nReturns KPIs for the given period (default: today in tenant timezone).",
                "operationId": "dashboardIndex",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Location-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Start date (defaults to today)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "End date (defaults to today)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dashboard KPIs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "period": {
                                            "properties": {
                                                "from": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "to": {
                                                    "type": "string",
                                                    "format": "date"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "kpis": {
                                            "properties": {
                                                "sale_count": {
                                                    "type": "integer"
                                                },
                                                "items_sold": {
                                                    "type": "integer"
                                                },
                                                "gross_revenue": {
                                                    "type": "number"
                                                },
                                                "returns_total": {
                                                    "type": "number"
                                                },
                                                "net_revenue": {
                                                    "type": "number"
                                                },
                                                "cogs": {
                                                    "type": "number"
                                                },
                                                "gross_profit": {
                                                    "type": "number"
                                                },
                                                "margin_percent": {
                                                    "type": "number"
                                                },
                                                "avg_ticket": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "stock_health": {
                                            "properties": {
                                                "low_stock": {
                                                    "type": "integer"
                                                },
                                                "out_of_stock": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "payment_breakdown": {
                                            "type": "object"
                                        },
                                        "top_items": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}